+++ kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/api/ApiPost.kt
@@ -1,0 +1,10 @@
+package com.gchristov.thecodinglove.kmpsearchdata.api
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ApiPost(
+    val title: String,
+    val url: String,
+    val imageUrl: String,
+)
