+++ kmp/kmp-search-data/src/commonMain/kotlin/com/gchristov/thecodinglove/kmpsearchdata/api/ApiSearchSession.kt
@@ -11,10 +11,3 @@ data class ApiSearchSession(
     val searchHistory: Map<String, List<Int>>,
     val currentPost: ApiPost?
 )
-
-@Serializable
-data class ApiPost(
-    val title: String,
-    val url: String,
-    val imageUrl: String,
-)
